Bug #1 - Wrong scene placement
The splash screen was placed at number 3, but because the first scene
has to be 0 (it was menu). Splash screen had object that menu required,
and that wont instantiate every time as we go to menu. So it was decided
to create it in splash screen and DontDestroyOnLoad(); was called.